Recall game, phase 1

+++ include/engine/system-config.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 namespace Engine
 {
   struct SystemConfig
@@ -7,8 +9,8 @@ namespace Engine
     // REQUIRED: modify this address to match the mac address of your PS3 controller
     static constexpr const char *macAddress = "00:1b:fb:8e:87:ac";
 
-    static constexpr const unsigned int numLeds = 300;
-    static constexpr const unsigned int serialBaud = 921600;
+    static constexpr const uint16_t numLeds = 300;
+    static constexpr const uint32_t serialBaud = 921600;
     static constexpr const unsigned short ledDimmerGpio = 34;
   };
 }
